5885: rename acceptmodalargs
