+++ src/apps/dashboard/dashboard.entry.tsx
@@ -10,7 +10,7 @@ import { ReservationMaterialDetailsProps } from "../../core/storybook/reservatio
 import { MaterialDetailsModalProps } from "../../core/storybook/materialDetailsModalArgs";
 import { GroupModalReservationsProps } from "../../core/storybook/reservationGroupModalArgs";
 import { DeleteReservationModalArgs } from "../../core/storybook/deleteReservationModalArgs";
-import { AcceptModalEntryTextProps } from "../../core/storybook/acceptModalArgs";
+import { AcceptFeesModalEntryTextProps } from "../../core/storybook/acceptFeesModalArgs";
 
 export interface DashBoardProps {
   // Url
@@ -58,7 +58,7 @@ const DashboardEntry: FC<
     GroupModalProps &
     GroupModalLoansProps &
     DeleteReservationModalArgs &
-    AcceptModalEntryTextProps &
+    AcceptFeesModalEntryTextProps &
     GroupModalReservationsProps &
     ReservationMaterialDetailsProps &
     MaterialDetailsModalProps
